third_party/WebKit/LayoutTests/fast/forms/date-multiple-fields/date-multiple-fields-ax-aria-attributes-expected.txt

This test checks accessible field labels in multiple fields date input UI.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Initial empty state
PASS focusedElementDescription() is "Month TestLabel, mm, TestTitle, intValue:0, range:1-12"
PASS focusedElementDescription() is "Day TestLabel, dd, TestTitle, intValue:0, range:1-31"
PASS focusedElementDescription() is "Year TestLabel, yyyy, TestTitle, intValue:0, range:1-275760"
Non-empty value
PASS focusedElementDescription() is "Month TestLabel, 10, TestTitle, intValue:10, range:1-12"
PASS focusedElementDescription() is "Day TestLabel, 09, TestTitle, intValue:9, range:1-31"
PASS focusedElementDescription() is "Year TestLabel, 2012, TestTitle, intValue:2012, range:1-275760"
Empty value
PASS focusedElementDescription() is "Year TestLabel, yyyy, TestTitle, intValue:0, range:1-275760"
PASS focusedElementDescription() is "Day TestLabel, dd, TestTitle, intValue:0, range:1-31"
PASS focusedElementDescription() is "Month TestLabel, mm, TestTitle, intValue:0, range:1-12"

PASS successfullyParsed is true

TEST COMPLETE
TestLabel
